% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_star_Intensity.R
\name{alpha_star_Intensity}
\alias{alpha_star_Intensity}
\title{alpha_star_Intensity}
\usage{
alpha_star_Intensity(Intensity, c0, c1, c2)
}
\arguments{
\item{Intensity}{A numerical value, the precipitation intensity, in mm/h}

\item{c0, c1, c2}{Model parameters}
}
\value{
A numerical value, the value of the scaled \eqn{\alpha}
}
\description{
Quadratic model of the scaled \eqn{\alpha} as a function of the precipitation intensity.
It returns the value of the scaled \eqn{\alpha} given the model parameters \eqn{c_0, c_1} and \eqn{c_2}
}
\references{
Maloku, K., Hingray, B., and Evin, G. (2023).
Accounting for precipitation asymmetry in a multiplicative random cascade disaggregation model.
Hydrol. Earth Syst. Sci., 27, 3643–3661, https://doi.org/10.5194/hess-27-3643-2023
}
\author{
Kaltrina Maloku
}
