% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_res_11.25_to_30.R
\name{change_res_11.25_to_30}
\alias{change_res_11.25_to_30}
\title{change_res_11.25_to_30}
\usage{
change_res_11.25_to_30(x)
}
\arguments{
\item{x}{A vector of numerical values of precipitation at a resolution of 11.25 minutes}
}
\value{
A vector of numerical values of precipitation at a resolution of 30 minutes
}
\description{
Convert a precipitation time series at a resolution of 11.25 minutes into a 30-minute time series, assuming a uniform distribution in the 11.25-minute interval
}
\author{
Kaltrina Maloku
}
