% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_res_22.5_to_60.R
\name{change_res_22.5_to_60}
\alias{change_res_22.5_to_60}
\title{change_res_22.5_to_60}
\usage{
change_res_22.5_to_60(x)
}
\arguments{
\item{x}{A vector of numerical values of precipitation at a resolution of 22.5 minutes}
}
\value{
A vector of numerical values of precipitation at a resolution of 60 minutes
}
\description{
Convert a precipitation time series at a resolution of 22.5 minutes into a 60-minute time series, assuming a uniform distribution in the 22.5-minute interval
}
\author{
Kaltrina Maloku
}
