% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_score.R
\name{compute.score}
\alias{compute.score}
\title{compute_score}
\usage{
compute.score(sim.mean, sim.sd, q05, q95, stat.obs)
}
\arguments{
\item{sim.mean}{A numerical value, the mean among simulated scenarios for a given statistic}

\item{sim.sd}{A numerical value, the standard deviation among simulated scenarios for a given statistic}

\item{q05}{A numerical value, the 5th percentile among simulated scenarios for a given statistic}

\item{q95}{A numerical value, the 95th percentile among simulated scenarios for a given statistic}

\item{stat.obs}{A numerical value, the statistic calculated from the observations}
}
\value{
A numerical value, the score as assigned by the CASE framework
}
\description{
A function that calculates the score as assigned by the CASE framework
}
\references{
Maloku, K., Hingray, B., and Evin, G. (2023).
Accounting for precipitation asymmetry in a multiplicative random cascade disaggregation model.
Hydrol. Earth Syst. Sci., 27, 3643–3661, https://doi.org/10.5194/hess-27-3643-2023
}
\author{
Kaltrina Maloku
}
