% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disaggregate_precip_MRC_Intensity.R
\name{disaggregate_precip_MRC_Intensity}
\alias{disaggregate_precip_MRC_Intensity}
\title{disaggregate_precip_MRC_Intensity}
\usage{
disaggregate_precip_MRC_Intensity(
  vecPrecip_target,
  vecDates_target,
  params_scaling,
  by_season,
  res_coarse_aggLevel,
  res_fine_aggLevel,
  asymmetry_option
)
}
\arguments{
\item{vecPrecip_target}{A vector of observed precipitation at coarse resolution, in mm}

\item{vecDates_target}{The corresponding vector of dates}

\item{params_scaling}{A data frame of parameters needed for disaggregation}

\item{by_season}{Logical. Are the parameters estimated on a seasonal basis? If F, the parameters are estimated on a monthly basis, if T on a seasonal basis}

\item{res_coarse_aggLevel}{Resolution of the time series to be disaggregated, in minutes}

\item{res_fine_aggLevel}{Target resolution of the disaggregated time series, in minutes}

\item{asymmetry_option}{Logical. Should the disaggregation be dependent on the asymmetry model?}
}
\value{
A vector of high resolution precipitation
}
\description{
A function that disaggregates a coarse resolution time series into a fine resolution time series according to the MRC model, 
where the dependence on precipitation intensity is taken into account
}
\references{
Maloku, K., Hingray, B., and Evin, G. (2023).
Accounting for precipitation asymmetry in a multiplicative random cascade disaggregation model.
Hydrol. Earth Syst. Sci., 27, 3643–3661, https://doi.org/10.5194/hess-27-3643-2023
}
\author{
Kaltrina Maloku
}
