% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_EGPD_IDF.R
\name{fit_EGPD_IDF}
\alias{fit_EGPD_IDF}
\title{fit_EGPD_IDF}
\usage{
fit_EGPD_IDF(
  vec.precip,
  vec.dates,
  durations = c(1, 2, 3, 6, 10, 12, 16, 18, 24, 48, 72),
  declustering_duration = c(1, 2, 3, 6, 10, 12, 16, 18, 24, 48, 72)
)
}
\arguments{
\item{vec.precip}{A vector of observed hourly precipitation}

\item{vec.dates}{The corresponding vector of dates}

\item{durations}{Durations in hours (for hourly data)}

\item{declustering_duration}{Declustering to reduce serial dependence description}
}
\value{
A list containing the EGPD parameters by month. Each element of the list is a matrix of size nStation x 3
}
\description{
A function that estimates the parameters of the EGPD
}
\author{
Kaltrina Maloku
}
