% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_IDF_model.R
\name{fit_IDF_model}
\alias{fit_IDF_model}
\title{fit_IDF_model}
\usage{
fit_IDF_model(station_data, durations, declustering_duration)
}
\arguments{
\item{station_data}{A matrix of precipitation data aggregated over different durations corresponding to one column}

\item{durations}{Durations in hours (for hourly data)}

\item{declustering_duration}{Declustering to reduce serial dependence description}
}
\value{
A list of parameters
}
\description{
fit_IDF_model
}
\author{
Abubakar Haruna
}
