% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_symm_beta_dist.R
\name{fit_symm_beta_dist}
\alias{fit_symm_beta_dist}
\title{fit_symm_beta_dist}
\usage{
fit_symm_beta_dist(x)
}
\arguments{
\item{x}{A vector of weights}
}
\value{
The estimated \eqn{\alpha} parameter
}
\description{
A function that fits a symmetric Beta distribution, with the probability density function:
\deqn{f(x, \alpha) = \frac{1}{B(\alpha, \alpha)} {x}^{(\alpha-1)}(1-x)^{(\alpha-1)}}
}
\details{
If sample size is less than 10, return NA
}
\references{
Maloku, K., Hingray, B., and Evin, G. (2023).
Accounting for precipitation asymmetry in a multiplicative random cascade disaggregation model.
Hydrol. Earth Syst. Sci., 27, 3643–3661, https://doi.org/10.5194/hess-27-3643-2023
}
\author{
Kaltrina Maloku
}
