% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_alpha2.R
\name{get_alpha2}
\alias{get_alpha2}
\title{get_alpha2}
\usage{
get_alpha2(alpha, m)
}
\arguments{
\item{alpha}{A numerical value, the parameter \eqn{\alpha} of the symmetric Beta distribution}

\item{m}{A numerical value, the mean of the Beta distribution}
}
\value{
A numerical value, the parameter \eqn{\alpha_{2}} of the Beta distribution
}
\description{
A function that estimates the parameter \eqn{\alpha_{2}} of the Beta distribution given the parameter \eqn{\alpha} of the symmetric Beta distribution and its mean
}
\references{
Maloku, K., Hingray, B., and Evin, G. (2023).
Accounting for precipitation asymmetry in a multiplicative random cascade disaggregation model.
Hydrol. Earth Syst. Sci., 27, 3643–3661, https://doi.org/10.5194/hess-27-3643-2023
}
\author{
Kaltrina Maloku
}
