% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_alpha_aggLevel.R
\name{get_alpha_aggLevel}
\alias{get_alpha_aggLevel}
\title{get_alpha_aggLevel}
\usage{
get_alpha_aggLevel(alpha0, H, res_aggLevel)
}
\arguments{
\item{alpha0, H}{Model parameters}

\item{res_aggLevel}{A numerical value, the current temporal aggregation level, in minutes}
}
\value{
A numerical value, the parameter \eqn{\alpha} of the Beta distribution
}
\description{
Alpha model as a function of the temporal aggregation level. 
It returns the value of the parameter \eqn{\alpha} of the Beta distribution given the model parameters \eqn{\alpha_{0}} and \eqn{H}
}
\references{
Maloku, K., Hingray, B., and Evin, G. (2023).
Accounting for precipitation asymmetry in a multiplicative random cascade disaggregation model.
Hydrol. Earth Syst. Sci., 27, 3643–3661, https://doi.org/10.5194/hess-27-3643-2023
}
\author{
Kaltrina Maloku
}
