% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_alpha_intensity_1par.R
\name{get_alpha_intensity_1par}
\alias{get_alpha_intensity_1par}
\title{get_alpha_intensity_1par}
\usage{
get_alpha_intensity_1par(Intensity, I_min, I_max, K)
}
\arguments{
\item{Intensity}{A numerical value, the precipitation intensity, in mm/h}

\item{I_min}{For intensities smaller than this value, the model is constant. Default 0.1 mm/h}

\item{I_max}{For intensities higher than this value, the model is constant. Default 10 mm/h}

\item{K}{Model parameter}
}
\value{
A numerical value, the parameter \eqn{\alpha} of the Beta distribution
}
\description{
Model of the parameter \eqn{\alpha} of the Beta distribution as a function of precipitation intensity.
Its single parameter \eqn{K} explains the degree of dependency to intensity.
The model is considered to be constant for intensities smaller than \code{I_min} and higher than \code{I_max}
}
\references{
Maloku, K., Hingray, B., and Evin, G. (2023).
Accounting for precipitation asymmetry in a multiplicative random cascade disaggregation model.
Hydrol. Earth Syst. Sci., 27, 3643–3661, https://doi.org/10.5194/hess-27-3643-2023
}
\author{
Kaltrina Maloku
}
