% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_col_aggLevel.R
\name{get_col_aggLevel}
\alias{get_col_aggLevel}
\title{get_col_aggLevel}
\usage{
get_col_aggLevel(aggLevel)
}
\arguments{
\item{aggLevel}{A vector providing the temporal aggregation levels, in minutes}
}
\value{
A vector of colors, of the same length as \code{aggLevel}
}
\description{
A function for plots
}
\author{
Kaltrina Maloku
}
