% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_indices_center_1280.R
\name{get_indices_center_1280}
\alias{get_indices_center_1280}
\title{get_indices_center_1280}
\usage{
get_indices_center_1280(nb_days)
}
\arguments{
\item{nb_days}{A numerical value, the number of days in the time series}
}
\value{
A vector of indices
}
\description{
A function that extracts "days" of 1280 minutes from a series of days of 1440 minutes
}
\references{
Maloku, K., Hingray, B., and Evin, G. (2023).
Accounting for precipitation asymmetry in a multiplicative random cascade disaggregation model.
Hydrol. Earth Syst. Sci., 27, 3643–3661, https://doi.org/10.5194/hess-27-3643-2023
}
\author{
Kaltrina Maloku
}
