% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_intensity_class_id.R
\name{get_intensity_class_id}
\alias{get_intensity_class_id}
\title{get_intensity_class_id}
\usage{
get_intensity_class_id(x, I_min = 0.001, res_aggLevel)
}
\arguments{
\item{x}{A vector of precipitation amounts, in mm}

\item{I_min}{A numerical value, the minimum precipitation observed, in mm/h. Default 0.001 mm/h for mean areal precipitation. For rain gauge data with a tipping volume of 0.1 mm, we suggest a value of 0.1*60/1440}

\item{res_aggLevel}{A numerical value, the current aggregation level, in minutes}
}
\value{
A vector of the same length as x, indicating the intensity class id of each precipitation amount
}
\description{
A function that, for a given precipitation amount, returns its intensity class id
}
\references{
Maloku, K., Hingray, B., and Evin, G. (2023).
Accounting for precipitation asymmetry in a multiplicative random cascade disaggregation model.
Hydrol. Earth Syst. Sci., 27, 3643–3661, https://doi.org/10.5194/hess-27-3643-2023
}
\author{
Kaltrina Maloku
}
