% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_length_spell_mean_sd.R
\name{get_length_spell_mean_sd}
\alias{get_length_spell_mean_sd}
\title{get_length_spell_mean_sd}
\usage{
get_length_spell_mean_sd(x)
}
\arguments{
\item{x}{A vector of precipitation, in mm}
}
\value{
A list containing the mean and standard deviation of the length of dry and wet spells
}
\description{
A function that calculates the mean and standard deviation of the length of dry and wet spells
}
\references{
Maloku, K., Hingray, B., and Evin, G. (2023).
Accounting for precipitation asymmetry in a multiplicative random cascade disaggregation model.
Hydrol. Earth Syst. Sci., 27, 3643–3661, https://doi.org/10.5194/hess-27-3643-2023
}
\author{
Kaltrina Maloku
}
