% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mean_z.R
\name{get_mean_z}
\alias{get_mean_z}
\title{get_mean_z}
\usage{
get_mean_z(z, lambda)
}
\arguments{
\item{z}{A vector of \eqn{z} index}

\item{lambda}{Model parameter}
}
\value{
A vector of the same length as \eqn{z}
}
\description{
Model of the mean of the cascade weights.
It returns the modelled mean of the distribution given the asymmetry \eqn{z} index and the model parameter \eqn{\lambda}
}
\references{
Maloku, K., Hingray, B., and Evin, G. (2023).
Accounting for precipitation asymmetry in a multiplicative random cascade disaggregation model.
Hydrol. Earth Syst. Sci., 27, 3643–3661, https://doi.org/10.5194/hess-27-3643-2023
}
\author{
Kaltrina Maloku
}
