% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_phi_z.R
\name{get_phi_z}
\alias{get_phi_z}
\title{get_phi_z}
\usage{
get_phi_z(z, nu)
}
\arguments{
\item{z}{A vector of \eqn{z} index}

\item{nu}{Model parameter}
}
\value{
A vector of the same length as \eqn{z}
}
\description{
Probability asymmetry ratio model.
It returns the ratio \eqn{\varphi = \frac{p_{01}}{p_{10}+p_{01}}} given the asymmetry \eqn{z} index and the model parameter \eqn{\nu}
}
\references{
Maloku, K., Hingray, B., and Evin, G. (2023).
Accounting for precipitation asymmetry in a multiplicative random cascade disaggregation model.
Hydrol. Earth Syst. Sci., 27, 3643–3661, https://doi.org/10.5194/hess-27-3643-2023
}
\author{
Kaltrina Maloku
}
