% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_portion_of_1.R
\name{get_portion_of_1}
\alias{get_portion_of_1}
\title{get_portion_of_1}
\usage{
get_portion_of_1(x)
}
\arguments{
\item{x}{A vector of weights}
}
\value{
The portion of weights equal to 1
}
\description{
A function that finds the portion of the cascade weights that is equal to 1
}
\details{
If sample size is less than 10, return NA
}
\references{
Maloku, K., Hingray, B., and Evin, G. (2023).
Accounting for precipitation asymmetry in a multiplicative random cascade disaggregation model.
Hydrol. Earth Syst. Sci., 27, 3643–3661, https://doi.org/10.5194/hess-27-3643-2023
}
\author{
Kaltrina Maloku
}
