% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Px_Intensity_aggLevel.R
\name{get_Px_Intensity_aggLevel}
\alias{get_Px_Intensity_aggLevel}
\title{get_Px_Intensity_aggLevel}
\usage{
get_Px_Intensity_aggLevel(
  Intensity,
  res_aggLevel,
  a_mu,
  b_mu,
  a_sigma,
  b_sigma
)
}
\arguments{
\item{Intensity}{A numerical value, the precipitation intensity, in mm/h}

\item{res_aggLevel}{A numerical value, the current aggregation level, in minutes}

\item{a_mu, b_mu, a_sigma, b_sigma}{Model parameters}
}
\value{
A numerical value, the probability to divide the precipitation intensity
}
\description{
Intermittency model as a function of the temporal aggregation level.
It returns the probability to divide the precipitation intensity given the model parameters \eqn{a_\mu, b_\mu, a_\sigma} and \eqn{b_\sigma}
}
\references{
Maloku, K., Hingray, B., and Evin, G. (2023).
Accounting for precipitation asymmetry in a multiplicative random cascade disaggregation model.
Hydrol. Earth Syst. Sci., 27, 3643–3661, https://doi.org/10.5194/hess-27-3643-2023
}
\author{
Kaltrina Maloku
}
