% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_scaling_MRC.R
\name{get_scaling_params_Intensity}
\alias{get_scaling_params_Intensity}
\title{get_scaling_params_Intensity}
\usage{
get_scaling_params_Intensity(
  vecPrecip,
  vecDates,
  resVecPrecip = 10,
  by_season = T,
  threshold = 0.8,
  threshold_alpha_int = 0.8,
  threshold_asymm = 0.8,
  aggLevels = c(80, 160, 320, 640, 1280)
)
}
\arguments{
\item{vecPrecip}{a vector of observed precipitations}

\item{vecDates}{a vector of dates}

\item{resVecPrecip}{resolution of time series in minutes}

\item{by_season}{should the parameter estimation must be done on seasonal basis}

\item{threshold}{for weight discarding in all estimation procedure, in mm. Default 0.8 mm.}

\item{threshold_alpha_int}{for discarding alpha estimated at small intensities, in mm. Default 0.8 mm.}

\item{threshold_asymm}{for discarding weights of small intensities for estimation of asymmetry parameters, in mm. Default 0.8 mm.}

\item{aggLevels}{aggregation levels in the cascade estimation procedure, in minutes}
}
\value{
a list, a data frame of scaling parameters and a list of plots for showing empirical estimates and fitted scaling models
}
\description{
Estimate scaling parameters of a MRC model, where the cascade weights are 
considered to be dependent on the intensity of precipitation that has to be disaggregated.
In the first step observed weights are calculated for each temporal aggregation level. 
Also at this stage, for each observed precipitation intensity, its index z is calculated.  
Second, MRC parameters are estimated by considering a dependency to 
temporal aggregation level, precipitation intensity class and asymmetry index class z. 
In the last step, scaling models are fitted on estimated MRC parameter by 
the method of linear least square. Plots showing estimated paramters and their scaling models 
are also returned
}
\author{
Kaltrina Maloku
}
