% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_standard_stats_extremes.R
\name{get_standard_stats_extremes}
\alias{get_standard_stats_extremes}
\title{get_standard_stats_extremes}
\usage{
get_standard_stats_extremes(df_data)
}
\arguments{
\item{df_data}{A data frame containing observed and disaggregated precipitation time series scenarios at a certain temporal resolution}
}
\value{
A data frame of statistics
}
\description{
A function that calculates statistics from precipitation time series scenarios
}
\references{
Maloku, K., Hingray, B., and Evin, G. (2023).
Accounting for precipitation asymmetry in a multiplicative random cascade disaggregation model.
Hydrol. Earth Syst. Sci., 27, 3643–3661, https://doi.org/10.5194/hess-27-3643-2023
}
\author{
Kaltrina Maloku
}
