% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_z_class_id.R
\name{get_z_class_id}
\alias{get_z_class_id}
\title{get_z_class_id}
\usage{
get_z_class_id(z, class_nb = 10)
}
\arguments{
\item{z}{A vector of \eqn{z} index}

\item{class_nb}{Number of classes}
}
\value{
A vector of the same length as \eqn{z}, indicating the class id of each \eqn{z} index
}
\description{
A function that, for a given \eqn{z} index, returns the id of its class
}
\references{
Maloku, K., Hingray, B., and Evin, G. (2023).
Accounting for precipitation asymmetry in a multiplicative random cascade disaggregation model.
Hydrol. Earth Syst. Sci., 27, 3643–3661, https://doi.org/10.5194/hess-27-3643-2023
}
\author{
Kaltrina Maloku
}
