% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeanArealPrecipData1440min.R
\docType{data}
\name{MeanArealPrecipData1440min}
\alias{MeanArealPrecipData1440min}
\title{Precipitation data set observed at 1440-minute resolution}
\format{
A data frame of 2 variables:
\describe{
  \item{obs}{Mean areal precipitation, in mm}
  \item{date}{Date, in POSIXct format}
}
}
\usage{
MeanArealPrecipData1440min
}
\description{
Example of a 17-year daily time series of mean areal precipitation observed in Switzerland
}
\keyword{datasets}
