% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeanArealPrecipData60min.R
\docType{data}
\name{MeanArealPrecipData60min}
\alias{MeanArealPrecipData60min}
\title{Precipitation data set observed at 60-minute resolution}
\format{
A data frame of 2 variables:
\describe{
  \item{obs}{Mean areal precipitation, in mm}
  \item{date}{Date, in POSIXct format}
}
}
\usage{
MeanArealPrecipData60min
}
\description{
Example of a 17-year hourly time series of mean areal precipitation observed in Switzerland
}
\keyword{datasets}
