% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/month2season.R
\name{month2season}
\alias{month2season}
\title{month2season}
\usage{
month2season(vecMonth)
}
\arguments{
\item{vecMonth}{A vector of months given as integers 1:12}
}
\description{
A function that transforms a vector of months to seasons
}
\author{
Guillaume Evin
}
