% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrecipData40min.R
\docType{data}
\name{PrecipData40min}
\alias{PrecipData40min}
\title{Precipitation data set observed at 40-minute resolution}
\format{
A data frame of 2 variables:
\describe{
  \item{obs}{Precipitation amount, in mm}
  \item{date}{Date, in POSIXct format}
}
}
\usage{
PrecipData40min
}
\description{
Example of a 40-year precipitation time series observed at a station in Switzerland with a resolution of 40 minutes
}
\references{
Maloku, K., Hingray, B., and Evin, G. (2023).
Accounting for precipitation asymmetry in a multiplicative random cascade disaggregation model.
Hydrol. Earth Syst. Sci., 27, 3643–3661, https://doi.org/10.5194/hess-27-3643-2023
}
\keyword{datasets}
