% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrecipData60min.R
\docType{data}
\name{PrecipData60min}
\alias{PrecipData60min}
\title{Precipitation data set observed at 60-minute resolution}
\format{
A data frame with 149016 rows and 2 variables:
\describe{
  \item{obs}{Mean areal precipitation, in mm}
  \item{date}{Date, in POSIXct format}
}
}
\usage{
PrecipData60min
}
\description{
Example of a 17-year hourly time series of mean areal precipitation observed in Switzerland
}
\keyword{datasets}
