% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prWet.R
\name{prWet}
\alias{prWet}
\title{prWet}
\usage{
prWet(x)
}
\arguments{
\item{x}{A vector of precipitation, in mm}
}
\value{
A numerical value, the proportion of wet steps
}
\description{
A function that calculates the proportion of wet steps in a time series
}
\references{
Maloku, K., Hingray, B., and Evin, G. (2023).
Accounting for precipitation asymmetry in a multiplicative random cascade disaggregation model.
Hydrol. Earth Syst. Sci., 27, 3643–3661, https://doi.org/10.5194/hess-27-3643-2023
}
\author{
Kaltrina Maloku
}
