% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq.emp.R
\name{qq.emp}
\alias{qq.emp}
\title{qq.emp}
\usage{
qq.emp(x, p, a = 0.44)
}
\arguments{
\item{x}{A vector of data}

\item{p}{A numerical value, the probability associated with the quantile to be calculated}

\item{a}{Parameter associated with the Gumbel distribution for probability fitting, based on a Gringorten approach}
}
\value{
A numerical value, the empirical quantile
}
\description{
A function that calculates an empirical quantile based on a Gumbel transformation
}
\references{
Maloku, K., Hingray, B., and Evin, G. (2023).
Accounting for precipitation asymmetry in a multiplicative random cascade disaggregation model.
Hydrol. Earth Syst. Sci., 27, 3643–3661, https://doi.org/10.5194/hess-27-3643-2023
}
\author{
Kaltrina Maloku
}
