% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_fixed_window.R
\name{sum_fixed_window}
\alias{sum_fixed_window}
\title{sum_fixed_window}
\usage{
sum_fixed_window(x, k)
}
\arguments{
\item{x}{A vector of precipitation amounts, in mm}

\item{k}{Length of the fixed window}
}
\value{
A vector of sums over the window of length k, in mm
}
\description{
A function that calculates a sum over a fixed window
}
\details{
Sums are assigned to the leftmost element in the window
}
\author{
Kaltrina Maloku
}
